pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptlab-cli"
version = "0.1.0"
description = "Automated testing for LLM prompts. Like pytest, but for prompts."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Vignesh Pai" }
]
keywords = ["llm", "testing", "prompts", "claude", "openai", "evaluation", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
promptlab = "promptlab.cli:main"

[project.urls]
Homepage = "https://github.com/vigp17/promptlab"
Repository = "https://github.com/vigp17/promptlab"
Issues = "https://github.com/vigp17/promptlab/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src/promptlab"]
